requirements.txt
cohere
praw
tqdm
python-dotenv
html2image
